refactor: start using new dir swapper in save manager
This commit removes the legacy swapping system inside SaveManager and
replaces some of it with DirSwapper. It also changes some of the return
types of DirSwapper to be more descriptive of what happened, mostly
returning `anyhow::Result<Option<()>>` instead of `anyhow::Result<()>`
(with a string error message for some non-serious issues).